+++ src/lib.rs
@@ -194,6 +194,34 @@
 //! # fn main() {}
 //! ```
 //!
+//! ## Hidden Fields
+//!
+//! You can hide fields by skipping their setters on the builder struct.
+//!
+//! - Opt-out &mdash; skip setters via `#[builder(setter(skip))]` on individual fields.
+//! - Opt-in &mdash; set `#[builder(setter(skip))]` on the whole struct
+//!   and enable individual setters via `#[builder(setter)]`.
+//!
+//! The types of skipped fields must implement `Default`.
+//!
+//! ```rust
+//! # #[macro_use]
+//! # extern crate derive_builder;
+//! #
+//! #[derive(Builder)]
+//! struct SetterOptOut {
+//!     setter_present: u32,
+//!     #[builder(setter(skip))]
+//!     setter_skipped: u32,
+//! }
+//! # fn main() {}
+//! ```
+//!
+//! Alternatively you can use the more verbose form:
+//!
+//! - `#[builder(setter(skip="true"))]`
+//! - `#[builder(setter(skip="false"))]`
+//!
 //! ## Setter Visibility
 //!
 //! Setters are public by default. You can precede your struct (or field) with `#[builder(public)]`
@@ -206,7 +234,7 @@
 //!
 //! Setter methods are named after their corresponding field by default.
 //!
-//! You can precede your struct (or field) with e.g. `#[builder(setter_prefix="xyz")` to change
+//! You can precede your struct (or field) with e.g. `#[builder(setter(prefix="xyz"))` to change
 //! the method name to `xyz_foo` if the field is named `foo`. Note that an underscore is included
 //! by default, since Rust favors snake case here.
 //!
@@ -341,9 +369,6 @@ fn builder_for_struct(ast: syn::MacroInput) -> quote::Tokens {
         setter_fns.push(derive_setter(f, &f_opts, &attrs));
         builder_fields.push(derive_builder_field(f, &f_opts, &attrs));
         initializers.push(derive_initializer(f, &f_opts));
-        // NOTE: Looking forward for computation in interpolation
-        // - https://github.com/dtolnay/quote/issues/10
-        // => `quote!(#{f.vis} ...)
     }
 
     let builder_vis = opts.builder_visibility();
@@ -425,7 +450,7 @@ fn derive_initializer(f: &syn::Field, opts: &FieldOptions) -> quote::Tokens {
         initializer
     } else {
         trace!("Fallback to default initializer for `{}`.", opts.field_name());
-        quote!( #ident: default::Default(), )
+        quote!( #ident: Default::default(), )
     }
 }